qa_rag_system/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qa-rag-system"
version = "0.1.0"
description = "RAG-based Q&A system for technical documentation with cited sources"
requires-python = ">=3.10.0,<4.0.0"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "LangChain Team"}
]

dependencies = [
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-chroma>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-classic>=0.3.0",
    "pinecone-client>=3.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.5.0",
    "streamlit>=1.38.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "requests>=2.32.0",
    "tiktoken>=0.8.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.13.0",
    "mypy>=1.18.0",
]

[project.urls]
Homepage = "https://github.com/langchain-ai/langchain"
Documentation = "https://docs.langchain.com"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

